Fix `setValueOnMove` state on `Select` not syncing between connected stores
