+++ examples/select-combobox-store/style.css
@@ -1,0 +1,1 @@
+@import url("../select-combobox/style.css");
