+++ examples/select-combobox-store/test.ts
@@ -1,0 +1,1 @@
+import "../select-combobox/test.js";
