+++ examples/select-combobox/readme.md
@@ -12,7 +12,7 @@ tags:
 
 <div data-description>
 
-Combining <a href="/components/select">Select</a> and <a href="/components/combobox">Combobox</a> to create a dropdown with a search field that can be used to filter items.
+Combining [Select](/components/select) and [Combobox](/components/combobox) to create a dropdown with a search field that can be used to filter items.
 
 </div>
