+++ examples/select-grid-store/style.css
@@ -1,0 +1,1 @@
+@import url("../select-grid/style.css");
