+++ examples/select-grid-store/test.ts
@@ -1,0 +1,1 @@
+import "../select-grid/test.js";
