--- examples/select-provider-combobox/list.ts
@@ -1,36 +1,0 @@
-export default [
-  "Apple",
-  "Bacon",
-  "Banana",
-  "Broccoli",
-  "Burger",
-  "Cake",
-  "Candy",
-  "Carrot",
-  "Cherry",
-  "Chocolate",
-  "Cookie",
-  "Cucumber",
-  "Donut",
-  "Fish",
-  "Fries",
-  "Grape",
-  "Green apple",
-  "Hot dog",
-  "Ice cream",
-  "Kiwi",
-  "Lemon",
-  "Lollipop",
-  "Onion",
-  "Orange",
-  "Pasta",
-  "Pineapple",
-  "Pizza",
-  "Potato",
-  "Salad",
-  "Sandwich",
-  "Steak",
-  "Strawberry",
-  "Tomato",
-  "Watermelon",
-];
