--- examples/select-provider-combobox/style.css
@@ -1,43 +1,0 @@
-@import url("../select/style.css");
-
-.popover {
-  @apply
-    pt-0
-  ;
-}
-
-.combobox-wrapper {
-  @apply
-    sticky
-    bg-inherit
-    top-0
-    pt-2
-    mb-2
-    w-fit
-  ;
-}
-
-.combobox {
-  @apply
-    h-10
-    rounded
-    border-none
-    px-4
-    w-full
-    text-base
-    bg-gray-150
-    text-black
-    hover:bg-gray-200
-    dark:bg-gray-850
-    dark:text-white
-    dark:hover:bg-gray-900
-    focus-visible:ariakit-outline-input
-  ;
-}
-
-.select-item {
-  @apply
-    scroll-mt-14
-    w-full
-  ;
-}
